# week8/dist/swagger.json
{
	"openapi": "3.0.0",
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"UserSignUpResponse": {
				"properties": {
					"userId": {
						"type": "number",
						"format": "double",
						"description": "생성된 사용자 ID",
						"example": 1
					},
					"preferCategory": {
						"items": {
							"type": "string"
						},
						"type": "array",
						"description": "선호 카테고리 이름 목록",
						"example": [
							"일식",
							"중식"
						]
					}
				},
				"required": [
					"userId",
					"preferCategory"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_UserSignUpResponse_": {
				"properties": {
					"resultType": {
						"type": "string",
						"enum": [
							"SUCCESS"
						],
						"nullable": false
					},
					"error": {
						"type": "number",
						"enum": [
							null
						],
						"nullable": true
					},
					"data": {
						"$ref": "#/components/schemas/UserSignUpResponse"
					}
				},
				"required": [
					"resultType",
					"error",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_null_": {
				"properties": {
					"resultType": {
						"type": "string",
						"enum": [
							"SUCCESS"
						],
						"nullable": false
					},
					"error": {
						"type": "number",
						"enum": [
							null
						],
						"nullable": true
					},
					"data": {
						"type": "number",
						"enum": [
							null
						],
						"nullable": true
					}
				},
				"required": [
					"resultType",
					"error",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UserSignUpRequest": {
				"properties": {
					"email": {
						"type": "string",
						"description": "유저 이메일 (로그인 시 사용)",
						"example": "test@example.com"
					},
					"name": {
						"type": "string",
						"description": "유저 이름",
						"example": "희주"
					},
					"gender": {
						"type": "string",
						"description": "성별",
						"example": "FEMALE"
					},
					"birth": {
						"type": "string",
						"format": "date-time",
						"description": "생년월일",
						"example": "2000-01-01T00:00:00.000Z"
					},
					"address": {
						"type": "string",
						"description": "주소",
						"example": "서울특별시 강북구"
					},
					"phoneNumber": {
						"type": "string",
						"description": "전화번호",
						"example": "01012345678"
					},
					"password": {
						"type": "string",
						"description": "비밀번호",
						"example": "qwer1234!"
					},
					"preferences": {
						"items": {
							"type": "number",
							"format": "double"
						},
						"type": "array",
						"description": "선호 음식 카테고리 ID 배열",
						"example": [
							1,
							2
						]
					}
				},
				"required": [
					"email",
					"name",
					"gender",
					"birth",
					"phoneNumber",
					"password",
					"preferences"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse__storeId-number__": {
				"properties": {
					"resultType": {
						"type": "string",
						"enum": [
							"SUCCESS"
						],
						"nullable": false
					},
					"error": {
						"type": "number",
						"enum": [
							null
						],
						"nullable": true
					},
					"data": {
						"properties": {
							"storeId": {
								"type": "number",
								"format": "double"
							}
						},
						"required": [
							"storeId"
						],
						"type": "object"
					}
				},
				"required": [
					"resultType",
					"error",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateStoreRequest": {
				"properties": {
					"categoryId": {
						"type": "number",
						"format": "double",
						"description": "가게 카테고리 ID",
						"example": 1
					},
					"regionId": {
						"type": "number",
						"format": "double",
						"description": "가게가 속한 지역 ID",
						"example": 3
					},
					"name": {
						"type": "string",
						"description": "가게 이름",
						"example": "맘스터치 성수점"
					},
					"address": {
						"type": "string",
						"description": "가게 주소",
						"example": "서울특별시 성동구 성수동 123-45"
					},
					"masterNum": {
						"type": "number",
						"format": "double",
						"description": "가게 전화번호",
						"example": "0212345678"
					}
				},
				"required": [
					"categoryId",
					"regionId",
					"name",
					"address",
					"masterNum"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ReviewItem": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double",
						"description": "리뷰 ID",
						"example": 1
					},
					"userName": {
						"type": "string",
						"description": "리뷰 작성자 이름",
						"example": "희주"
					},
					"star": {
						"type": "number",
						"format": "double",
						"description": "별점",
						"example": 5
					},
					"content": {
						"type": "string",
						"description": "리뷰 내용",
						"example": "맛있어요!"
					}
				},
				"required": [
					"id",
					"userName",
					"star",
					"content"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ReviewListResponse": {
				"properties": {
					"data": {
						"items": {
							"$ref": "#/components/schemas/ReviewItem"
						},
						"type": "array",
						"description": "리뷰 목록"
					},
					"pagination": {
						"properties": {
							"cursor": {
								"type": "number",
								"format": "double",
								"nullable": true,
								"description": "다음 페이지 조회용 커서",
								"example": 15
							}
						},
						"required": [
							"cursor"
						],
						"type": "object"
					}
				},
				"required": [
					"data",
					"pagination"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_ReviewListResponse_": {
				"properties": {
					"resultType": {
						"type": "string",
						"enum": [
							"SUCCESS"
						],
						"nullable": false
					},
					"error": {
						"type": "number",
						"enum": [
							null
						],
						"nullable": true
					},
					"data": {
						"$ref": "#/components/schemas/ReviewListResponse"
					}
				},
				"required": [
					"resultType",
					"error",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse__reviewId-number__": {
				"properties": {
					"resultType": {
						"type": "string",
						"enum": [
							"SUCCESS"
						],
						"nullable": false
					},
					"error": {
						"type": "number",
						"enum": [
							null
						],
						"nullable": true
					},
					"data": {
						"properties": {
							"reviewId": {
								"type": "number",
								"format": "double"
							}
						},
						"required": [
							"reviewId"
						],
						"type": "object"
					}
				},
				"required": [
					"resultType",
					"error",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateReviewRequest": {
				"properties": {
					"userId": {
						"type": "number",
						"format": "double",
						"description": "리뷰 작성 사용자 ID",
						"example": 1
					},
					"storeId": {
						"type": "number",
						"format": "double",
						"description": "리뷰를 작성할 가게 ID",
						"example": 3
					},
					"star": {
						"type": "number",
						"format": "double",
						"description": "별점",
						"example": 5
					},
					"content": {
						"type": "string",
						"description": "리뷰 내용",
						"example": "치킨이 맛있어요!"
					}
				},
				"required": [
					"userId",
					"storeId",
					"star",
					"content"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_any_": {
				"properties": {
					"resultType": {
						"type": "string",
						"enum": [
							"SUCCESS"
						],
						"nullable": false
					},
					"error": {
						"type": "number",
						"enum": [
							null
						],
						"nullable": true
					},
					"data": {}
				},
				"required": [
					"resultType",
					"error",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse__missionId-number__": {
				"properties": {
					"resultType": {
						"type": "string",
						"enum": [
							"SUCCESS"
						],
						"nullable": false
					},
					"error": {
						"type": "number",
						"enum": [
							null
						],
						"nullable": true
					},
					"data": {
						"properties": {
							"missionId": {
								"type": "number",
								"format": "double"
							}
						},
						"required": [
							"missionId"
						],
						"type": "object"
					}
				},
				"required": [
					"resultType",
					"error",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateMissionRequest": {
				"properties": {
					"storeId": {
						"type": "number",
						"format": "double",
						"description": "미션이 등록될 가게 ID",
						"example": 1
					},
					"deadline": {
						"type": "string",
						"format": "date-time",
						"description": "미션 마감일",
						"example": "2026-12-31T23:59:59.000Z"
					},
					"condition": {
						"type": "string",
						"description": "미션 수행 조건",
						"example": "5000원 이상 주문하기"
					},
					"point": {
						"type": "number",
						"format": "double",
						"description": "미션 성공 시 지급 포인트",
						"example": 500
					}
				},
				"required": [
					"storeId",
					"deadline",
					"condition",
					"point"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse__userMissionId-number__": {
				"properties": {
					"resultType": {
						"type": "string",
						"enum": [
							"SUCCESS"
						],
						"nullable": false
					},
					"error": {
						"type": "number",
						"enum": [
							null
						],
						"nullable": true
					},
					"data": {
						"properties": {
							"userMissionId": {
								"type": "number",
								"format": "double"
							}
						},
						"required": [
							"userMissionId"
						],
						"type": "object"
					}
				},
				"required": [
					"resultType",
					"error",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ChallengeMissionRequest": {
				"properties": {
					"missionId": {
						"type": "number",
						"format": "double",
						"description": "도전할 미션 ID",
						"example": 2
					}
				},
				"required": [
					"missionId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse__userMissionId-number--completed-boolean__": {
				"properties": {
					"resultType": {
						"type": "string",
						"enum": [
							"SUCCESS"
						],
						"nullable": false
					},
					"error": {
						"type": "number",
						"enum": [
							null
						],
						"nullable": true
					},
					"data": {
						"properties": {
							"completed": {
								"type": "boolean"
							},
							"userMissionId": {
								"type": "number",
								"format": "double"
							}
						},
						"required": [
							"completed",
							"userMissionId"
						],
						"type": "object"
					}
				},
				"required": [
					"resultType",
					"error",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"OngoingMissionItem": {
				"properties": {
					"userMissionId": {
						"type": "number",
						"format": "double",
						"description": "사용자 미션 ID",
						"example": 1
					},
					"point": {
						"type": "number",
						"format": "double",
						"description": "획득 가능 포인트",
						"example": 500
					},
					"storeName": {
						"type": "string",
						"description": "가게 이름",
						"example": "버거킹 성수점"
					},
					"condition": {
						"type": "string",
						"description": "미션 조건",
						"example": "세트 메뉴 주문하기"
					}
				},
				"required": [
					"userMissionId",
					"point",
					"storeName",
					"condition"
				],
				"type": "object",
				"additionalProperties": false
			},
			"OngoingMissionListResponse": {
				"properties": {
					"data": {
						"items": {
							"$ref": "#/components/schemas/OngoingMissionItem"
						},
						"type": "array",
						"description": "진행 중인 미션 목록"
					}
				},
				"required": [
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_OngoingMissionListResponse_": {
				"properties": {
					"resultType": {
						"type": "string",
						"enum": [
							"SUCCESS"
						],
						"nullable": false
					},
					"error": {
						"type": "number",
						"enum": [
							null
						],
						"nullable": true
					},
					"data": {
						"$ref": "#/components/schemas/OngoingMissionListResponse"
					}
				},
				"required": [
					"resultType",
					"error",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"StoreMissionItem": {
				"properties": {
					"missionId": {
						"type": "number",
						"format": "double",
						"description": "미션 ID",
						"example": 1
					},
					"storeName": {
						"type": "string",
						"description": "가게 이름",
						"example": "교촌치킨 성수점"
					},
					"storeCategory": {
						"type": "string",
						"description": "가게 카테고리",
						"example": "패스트푸드"
					},
					"condition": {
						"type": "string",
						"description": "미션 조건",
						"example": "2만원 이상 주문하기"
					},
					"point": {
						"type": "number",
						"format": "double",
						"description": "보상 포인트",
						"example": 1000
					},
					"deadline": {
						"type": "string",
						"format": "date-time",
						"description": "미션 마감일",
						"example": "2026-06-30T23:59:59.000Z"
					}
				},
				"required": [
					"missionId",
					"storeName",
					"storeCategory",
					"condition",
					"point",
					"deadline"
				],
				"type": "object",
				"additionalProperties": false
			},
			"StoreMissionListResponse": {
				"properties": {
					"data": {
						"items": {
							"$ref": "#/components/schemas/StoreMissionItem"
						},
						"type": "array",
						"description": "가게 미션 목록"
					}
				},
				"required": [
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_StoreMissionListResponse_": {
				"properties": {
					"resultType": {
						"type": "string",
						"enum": [
							"SUCCESS"
						],
						"nullable": false
					},
					"error": {
						"type": "number",
						"enum": [
							null
						],
						"nullable": true
					},
					"data": {
						"$ref": "#/components/schemas/StoreMissionListResponse"
					}
				},
				"required": [
					"resultType",
					"error",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "playground-umc-7th-nodejs",
		"version": "1.0.0",
		"description": "UMC 테스트 프로젝트",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
		"/users/signup": {
			"post": {
				"operationId": "HandleUserSignUp",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_UserSignUpResponse_"
								}
							}
						}
					},
					"400": {
						"description": "중복된 이메일 에러",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_null_"
								}
							}
						}
					}
				},
				"description": "회원가입 API",
				"summary": "사용자의 회원가입을 처리합니다.",
				"tags": [
					"Users"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UserSignUpRequest"
							}
						}
					}
				}
			}
		},
		"/stores": {
			"post": {
				"operationId": "CreateStore",
				"responses": {
					"201": {
						"description": "가게 생성 성공",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse__storeId-number__"
								}
							}
						}
					},
					"400": {
						"description": "잘못된 요청",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_null_"
								}
							}
						}
					}
				},
				"description": "가게 생성 API",
				"summary": "새로운 가게를 생성합니다.",
				"tags": [
					"Store"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateStoreRequest"
							}
						}
					}
				}
			}
		},
		"/stores/{storeId}/reviews": {
			"get": {
				"operationId": "ListStoreReviews",
				"responses": {
					"200": {
						"description": "가게 리뷰 목록 조회 성공",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_ReviewListResponse_"
								}
							}
						}
					},
					"404": {
						"description": "가게를 찾을 수 없음",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_null_"
								}
							}
						}
					}
				},
				"description": "가게 리뷰 목록 조회 API",
				"summary": "특정 가게의 리뷰 목록을 조회합니다.",
				"tags": [
					"Store"
				],
				"security": [],
				"parameters": [
					{
						"description": "리뷰를 조회할 가게 ID",
						"in": "path",
						"name": "storeId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "커서 기반 페이지네이션 값",
						"in": "query",
						"name": "cursor",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/reviews": {
			"post": {
				"operationId": "CreateReview",
				"responses": {
					"201": {
						"description": "리뷰 생성 성공",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse__reviewId-number__"
								}
							}
						}
					},
					"400": {
						"description": "잘못된 요청",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_null_"
								}
							}
						}
					}
				},
				"description": "리뷰 생성 API",
				"summary": "리뷰를 작성합니다.",
				"tags": [
					"Review"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateReviewRequest"
							}
						}
					}
				}
			}
		},
		"/reviews/users/{userId}": {
			"get": {
				"operationId": "ListMyReviews",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_any_"
								}
							}
						}
					},
					"404": {
						"description": "사용자를 찾을 수 없음",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_null_"
								}
							}
						}
					}
				},
				"description": "내가 작성한 리뷰 조회 API",
				"summary": "특정 사용자의 리뷰 목록을 조회합니다.",
				"tags": [
					"Review"
				],
				"security": [],
				"parameters": [
					{
						"description": "리뷰를 조회할 사용자 ID",
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "커서 기반 페이지네이션 값",
						"in": "query",
						"name": "cursor",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/missions": {
			"post": {
				"operationId": "CreateMission",
				"responses": {
					"201": {
						"description": "미션 생성 성공",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse__missionId-number__"
								}
							}
						}
					},
					"400": {
						"description": "잘못된 요청",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_null_"
								}
							}
						}
					}
				},
				"description": "미션 생성 API",
				"summary": "새로운 미션을 생성합니다.",
				"tags": [
					"Mission"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateMissionRequest"
							}
						}
					}
				}
			}
		},
		"/missions/challenge/{userId}": {
			"post": {
				"operationId": "ChallengeMission",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse__userMissionId-number__"
								}
							}
						}
					},
					"400": {
						"description": "이미 진행 중인 미션",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_null_"
								}
							}
						}
					},
					"404": {
						"description": "미션 또는 사용자를 찾을 수 없음",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_null_"
								}
							}
						}
					}
				},
				"description": "미션 도전 API",
				"summary": "사용자가 미션에 도전합니다.",
				"tags": [
					"Mission"
				],
				"security": [],
				"parameters": [
					{
						"description": "미션에 도전할 사용자 ID",
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ChallengeMissionRequest"
							}
						}
					}
				}
			}
		},
		"/missions/complete/{userId}/{userMissionId}": {
			"post": {
				"operationId": "CompleteMission",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse__userMissionId-number--completed-boolean__"
								}
							}
						}
					},
					"404": {
						"description": "진행 중인 미션을 찾을 수 없음",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_null_"
								}
							}
						}
					}
				},
				"description": "미션 완료 API",
				"summary": "진행 중인 미션을 완료 처리합니다.",
				"tags": [
					"Mission"
				],
				"security": [],
				"parameters": [
					{
						"description": "미션을 완료할 사용자 ID",
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "진행 중인 사용자 미션 ID",
						"in": "path",
						"name": "userMissionId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/missions/ongoing/{userId}": {
			"get": {
				"operationId": "ListOngoingMissions",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_OngoingMissionListResponse_"
								}
							}
						}
					},
					"404": {
						"description": "사용자를 찾을 수 없음",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_null_"
								}
							}
						}
					}
				},
				"description": "진행 중인 미션 조회 API",
				"summary": "사용자의 진행 중인 미션 목록을 조회합니다.",
				"tags": [
					"Mission"
				],
				"security": [],
				"parameters": [
					{
						"description": "조회할 사용자 ID",
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/missions/store/{storeId}": {
			"get": {
				"operationId": "ListStoreMissions",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_StoreMissionListResponse_"
								}
							}
						}
					},
					"404": {
						"description": "가게를 찾을 수 없음",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_null_"
								}
							}
						}
					}
				},
				"description": "가게 미션 목록 조회 API",
				"summary": "특정 가게의 미션 목록을 조회합니다.",
				"tags": [
					"Mission"
				],
				"security": [],
				"parameters": [
					{
						"description": "미션 목록을 조회할 가게 ID",
						"in": "path",
						"name": "storeId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/api"
		}
	]
}